+++ src/main/java/roomescape/WebConfig.java
@@ -7,18 +7,15 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import roomescape.admin.AdminInterceptor;
 import roomescape.login.LoginMemberArgumentResolver;
 import roomescape.member.MemberService;
-import roomescape.util.JwtUtil;
 
 import java.util.List;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
     private final MemberService memberService;
-    private final JwtUtil jwtUtil;
 
-    public WebConfig(MemberService memberService, JwtUtil jwtUtil) {
+    public WebConfig(MemberService memberService) {
         this.memberService = memberService;
-        this.jwtUtil = jwtUtil;
     }
 
     @Override
@@ -28,7 +25,7 @@ public class WebConfig implements WebMvcConfigurer {
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new AdminInterceptor(jwtUtil))
+        registry.addInterceptor(new AdminInterceptor(memberService))
                 .addPathPatterns("/admin/**");
     }
 }
