--- src/main/java/roomescape/util/JwtUtil.java
@@ -1,37 +1,0 @@
-package roomescape.util;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-import roomescape.member.Member;
-
-import java.nio.charset.StandardCharsets;
-
-@Component
-public class JwtUtil {
-    private final String secretKey;
-
-    public JwtUtil(@Value("${roomescape.auth.jwt.secret}") String secretKey) {
-        this.secretKey = secretKey;
-    }
-
-    public Claims parseClaims(String token) {
-        return Jwts.parserBuilder()
-                .setSigningKey(Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8)))
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-    }
-
-    public String generateToken(Member member) {
-        return Jwts.builder()
-                .setSubject(member.getId().toString())
-                .claim("id", member.getId())
-                .claim("name", member.getName())
-                .claim("role", member.getRole())
-                .signWith(io.jsonwebtoken.security.Keys.hmacShaKeyFor(secretKey.getBytes()))
-                .compact();
-    }
-}
